requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.4.0
autobahn==25.9.1
Automat==25.4.16
billiard==4.2.1
black==25.1.0
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==45.0.7
daphne==4.0.0
dj-database-url==3.0.1
Django==5.2.6
django-autoslug==1.9.9
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-ninja==1.4.3
django-prometheus==2.4.1
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
dnspython==2.8.0
email-validator==2.3.0
fcm-django==2.3.1
firebase_admin==7.1.0
flower==2.0.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hiredis==3.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.13.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
kombu==5.5.4
msgpack==1.1.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.13
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybreaker==1.4.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==5.2.1
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
swapper==1.4.0
tornado==6.5.2
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.10.0
wsproto==1.2.0
zope.interface==8.0.1
